refactor: DTO 패키지 의존성 통일
