+++ backend/src/main/java/kr/touroot/member/controller/MyPageController.java
@@ -11,9 +11,9 @@ import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
 import kr.touroot.global.auth.dto.MemberAuth;
 import kr.touroot.global.exception.dto.ExceptionResponse;
-import kr.touroot.member.dto.MyTravelogueResponse;
-import kr.touroot.member.dto.ProfileResponse;
 import kr.touroot.member.dto.request.ProfileUpdateRequest;
+import kr.touroot.member.dto.response.MyTravelogueResponse;
+import kr.touroot.member.dto.response.ProfileResponse;
 import kr.touroot.member.service.MyPageFacadeService;
 import kr.touroot.travelplan.dto.response.PlanResponse;
 import lombok.RequiredArgsConstructor;
