+++ backend/src/main/java/kr/touroot/member/service/MemberService.java
@@ -4,9 +4,9 @@ import kr.touroot.authentication.infrastructure.PasswordEncryptor;
 import kr.touroot.global.auth.dto.MemberAuth;
 import kr.touroot.global.exception.BadRequestException;
 import kr.touroot.member.domain.Member;
-import kr.touroot.member.dto.ProfileResponse;
 import kr.touroot.member.dto.request.MemberRequest;
 import kr.touroot.member.dto.request.ProfileUpdateRequest;
+import kr.touroot.member.dto.response.ProfileResponse;
 import kr.touroot.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
